Get rid of MAX_GETBLOCKTXN_DEPTH constant
Instead of allowing getblocktxn requests for the most recent 10 blocks (and
ignoring requests for other blocks), allow getblocktxn requests for blocks more
recent than the oldest outstanding cmpctblock message.

This fixes a race condition that can happen during tests where more than 10
cmpctblock messages are sent to a peer before the first incoming getblocktxn
messages can be processed, causing sync timeouts when the peer is not to be
able to get in sync.

This is an alternative to the fix in https://github.com/chaincodelabs/ccl-bitcoin/commit/fix-8842-sync-timeouts

Disadvantages of this implementation:

- Requires a new state member.

- Larger code change than the other fix.

Advantages of this implementation:

- Gets rid of the MAX_GETBLOCKTXN_DEPTH constant (which is just arbitrarily set
  at twice the value of MAX_CMPCTBLOCK_DEPTH).

- Makes the protocol simpler and more deterministic from the perspective of
  receiving peers. Instead of getblocktxn responses yielding either blocktxn or
  block messages depending on timing and unrelated internal state of the
  sending node, valid getblocktxn requests will reliably trigger blocktxn
  responses and no other types of response.

- Unlike the other fix, this doesn't require an update to BIP-152.